Added additional solution to problem 3
Added another solution to problem 3 that uses the fact that any integer
can be written as a product of prime factors (i.e. the fundamental
theorem of arithmetic).

Currently this solution checks all odd numbers greater than or equal to
3, it may be more efficient to only check prime numbers (although this
would require actually generating the next prime number which may be
more expensive computationally).

Also, the code will run correctly for the integer given in the problem
(600851475143), however, some changes would be needed in order to use
the code in the general case (for all integers). In particular, it would
be necessary to check if the initial integer is even.